refactor: Separate the sources
